Refactor Pokemon-evolutions into seperate files, (gen 1)
