+++ src/data/balance/pokemon-evolutions/pokemon-evolution-alt-levels.ts
@@ -1,0 +1,27 @@
+/** Used for most item based evolutions. The 36 comes from the fact that it's a good
+ * level for a fully evolved Pokemon and also because that's what Pokemon Quest used
+ */
+export const GENERIC_ITEM_EVO_LEVEL = 36;
+/**
+ * Used for rarer evolutionary items or when the base Pokemon
+ * is already pretty strong and it would not make sense
+ * for it to GENERIC_ITEM_EVO_LEVEL
+ */
+export const ADVANCED_ITEM_EVO_LEVEL = 45;
+/**
+ * Baby Pokemon are significantly weaker so there are two tiers
+ * of expected level for when they should appear as enemies
+ */
+export const BABY_HAPPINESS_EVO_LEVEL = 10;
+export const HAPPINESS_EVO_LEVEL = 25;
+/** Filler value for now, break down into individual Pokemon */
+export const KNOW_MOVE_EVO_LEVEL = 36;
+/** Learns Rage Fist at level 35 in gen 9 */
+export const ANNIHILAPE_LEVEL = 35;
+/** Learns Ancient Power at level 24 in gen 8 */
+export const TANGROWTH_LEVEL = 24;
+/** Learns rollout at level 6 in gen 8 and level 33 in gen 7 */
+export const LICKILICKY_LEVEL = 33;
+/** Learns mimic at level 32 in gen 8 and 15 in gen 7 */
+export const MR_MIME_LEVEL = 15;
+export const SUPER_LATE = 64;
