+++ src/data/balance/trainer-configs/rival-trainer-configs.ts
@@ -1,4 +1,4 @@
-import { pokemonEvolutions, pokemonPrevolutions } from "#app/data/balance/pokemon-evolutions";
+import { pokemonEvolutions, pokemonPrevolutions } from "../pokemon-evolutions/init-pokemon-evolutions";
 import type PokemonSpecies from "#app/data/pokemon-species";
 import {
   getRandomPartyMemberFunc,
