+++ src/data/pokemon-species.ts
@@ -6,8 +6,8 @@ import i18next from "i18next";
 import type { GameMode } from "#app/game-mode";
 import { randSeedInt, randSeedGauss } from "#app/utils";
 import type { GrowthRate } from "#enums/growth-rates";
-import type { EvolutionLevel } from "#app/data/balance/pokemon-evolutions";
-import { pokemonEvolutions, pokemonPrevolutions } from "#app/data/balance/pokemon-evolutions";
+import type { EvolutionLevel } from "#app/data/balance/pokemon-evolutions/pokemon-evolutions";
+import { pokemonEvolutions, pokemonPrevolutions } from "./balance/pokemon-evolutions/init-pokemon-evolutions";
 import type { ElementalType } from "#enums/elemental-type";
 import { variantData } from "#app/data/variant";
 import { SpeciesFormKey } from "#enums/species-form-key";
